requirements.txt
python-dotenv==1.0.1
psycopg2-binary==2.9.9
faker==24.14.1